Move DatabaseHandler.java to api package
