+++ app/src/main/java/za/co/social_engineer/www/socialengineer/SEADMActivity.java
@@ -7,6 +7,7 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
 
+import za.co.social_engineer.www.socialengineer.api.DatabaseHandler;
 import za.co.social_engineer.www.socialengineer.model.Question;
 
 public class SEADMActivity extends AppCompatActivity {
